feat: implement Melhor Envio shipment generation + fix volume weight unit
- Add MelhorEnvioShipmentService: full 3-step flow (cart→checkout→generate)
- Fix VolumeInfo.weight: int→double, weight in kg (not grams)
- Fix buildVolumes: remove kg-to-grams conversion, API expects kg
- Add protocol column to shipments table (V9 migration)
- Fix guest address storage in OrderFacade for shipment generation
- Add store config properties for Melhor Envio sender info
- Fix AsaasWebhookService: call generateShipment after order approved
- All 45 tests pass, build succeeds

+++ src/main/java/com/lisory/backend/envios/melhorenvio/dto/MelhorEnvioCartRequest.java
@@ -39,7 +39,7 @@ public record MelhorEnvioCartRequest(
         @JsonProperty("height") int height,
         @JsonProperty("width") int width,
         @JsonProperty("length") int length,
-        @JsonProperty("weight") int weight
+        @JsonProperty("weight") double weight
     ) {}
 
     public record CartOptions(
